MSP430 firmware version 4.x
This release adds support for OTA updating
This branch is used to build the firmware that is to be sent over the
air.  While this can be manually flashed to a device, it is recommended
to be send through OTA.  When updating through OTA it is necessary that a
proper base code version is available inside the device. The base
firmware will act as a fall back if there are issues in the OTA update
process.

Detailed version numbers:
v4.0 - TI  compiler, New charger
v4.1 - GNU compiler, New charger
v4.2 - TI  compiler, Old charger
v4.3 - GNU compiler, Old charger